+++ action_server/docs/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## Unreleased
+## 0.0.24 - 2024-02-23
 
 - Properly use all lines from docstring description to feed to the `openapi.json`.
+- When creating project from template, don't create a root directory.
 
 ## 0.0.23 - 2024-02-23